feat: configure ssh and zsh for better development experience
This commit configures SSH for easier key management and updates
Zsh to set environment variables based on the OS. It adds
`.ssh/config` and `.ssh/config.d/.gitignore` for SSH
configuration and ignores in the `.ssh/config.d` directory. It
also updates `.zprofile` and `.zshrc` to manage environment
variables for different operating systems.